🧹 Move theme utilities to application theme directory
- Move theme-to-css-variables.ts → themes/application/
- Move theme-adapter.ts → themes/application/
- Update all imports to use local paths

These utilities are only used by the application theme system (styleguide),
so they belong with those themes rather than in lib/
