+++ themes/application/index.ts
@@ -15,7 +15,7 @@ export {
 } from './theme-registry'
 
 // Export theme creation utility
-export { createTheme } from '@/lib/theme-adapter'
+export { createTheme } from './theme-adapter'
 
 // For backward compatibility and convenience
 import { getAllThemes } from './theme-registry'
