+++ themes/application/theme-to-css-variables.ts
